+++ docs/source/changelog.rst
@@ -12,6 +12,7 @@ Version (dev)
 * Treat ``ogg`` files as valid sounds in various places.
 * Consistently raise :class:`TypeError` if invalid types are sent to a :class:`Tokenizer`.
 * Properly add ``base_plat_train`` class resource function and fix the ``item_teamflag``.
+* :bee-pack-issue:`4299`: Fix sprites and decals not being detected properly for packing.
 
 -------------
 Version 2.5.0